feat(schedule): Implement basic types needed for schedules - Added CustomWeekdays and CustomWeekdaysVec for fine grained control - Added radio input for Options - Added DEV based environment
